requirements.txt
python-fasthtml
fasthtml-hf
pillow
